Moved interfaces from controller and service package to stereotypes in each one of them
