+++ src/main/java/fr/motormingle/api/controller/MotorbikeController.java
@@ -1,5 +1,6 @@
 package fr.motormingle.api.controller;
 
+import fr.motormingle.api.controller.stereotype.GetController;
 import fr.motormingle.api.entity.Motorbike;
 import fr.motormingle.api.service.MotorbikeService;
 import org.springframework.beans.factory.annotation.Autowired;
