+++ src/main/java/fr/motormingle/api/controller/stereotype/GetController.java
@@ -1,4 +1,4 @@
-package fr.motormingle.api.controller;
+package fr.motormingle.api.controller.stereotype;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
