+++ src/main/java/fr/motormingle/api/service/MotorbikeService.java
@@ -3,6 +3,7 @@ package fr.motormingle.api.service;
 import fr.motormingle.api.entity.Motorbike;
 import fr.motormingle.api.exception.NotFoundException;
 import fr.motormingle.api.repository.MotorbikeRepository;
+import fr.motormingle.api.service.stereotype.FindService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
