+++ src/main/java/fr/motormingle/api/service/PhotoService.java
@@ -4,6 +4,7 @@ import fr.motormingle.api.entity.Photo;
 import fr.motormingle.api.entity.PhotoId;
 import fr.motormingle.api.exception.NotFoundException;
 import fr.motormingle.api.repository.PhotoRepository;
+import fr.motormingle.api.service.stereotype.FindService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
